fix: Make several improvements to countries table
It makes sense to adhere to the ISO ALPHA-3 spec for country codes, which can be up to 3 characters in length.

The us_state_dept_recognized field has several potential uses, such as whether (when operating within the US) we're permitted to do business with an individual residing in a specific country (certain types of sanctions may preclude doing so).

The pay_pal_supported field is required to determine whether or not we're able to send artist payments via PayPal.