fix: Flip relationship between Song and FlacFile
Previously, a FlacFile belonged to a Song, but we decided that this is actually
"backwards" if we want to allow for the same FlacFile to be associated with
more than one Song (which we do).

Real-world use-cases for this include a Song being released as a "single", and
later released on a full-length album. With this new approach, we can "link"
the two tracks behind-the-scenes, so that fixing/updating the Song on one Album
will affect all instances (and prevent wasteful duplication on the filesystem).