tests: Go back to using MySQL for tests
SQLite has a number of problems, not least of which is that foreign
key constraints are disabled by default, with no easy nor reliable
means by which to enable them in all contexts.

Furthermore, it makes the most sense to test against the same
database that is used in production.

With some tuning, we should be able to achieve nearly comparable
results with MySQL.