tests: Replace RefreshDatabase trait with DatabaseTransactions trait
Switching back to MySQL (from SQLite) for testing caused a huge
performance hit, and this change is in an effort to reduce it.

Also, auto-incrementing values on InnoDB tables are not rolled-back
when a transaction is rolled-back, so they keep increasing across
tests. Accordingly, I removed the findById() test from
RepositoryReadonlyTestCase, if and until we decide to replace it with
something that is useful, given this change to DataBaseTransactions.

Finally, the larger change here has implications for how the database
is migrated (see base TestCase). Currently, we're seeding "static"
tables that don't change, e.g., "countries" and "genres", once at the
beginning of the test-run, but maybe we should eliminate that, too, and
seed them only for tests that need them (which is most tests, FWIW).