fix: Iterate over related models to force deletion events to fire
The manual explains why this is necessary:

"When executing a mass delete statement via Eloquent, the deleting and
deleted model events will not be fired for the deleted models. This is
because the models are never actually retrieved when executing the
delete statement."

In effect, this means that deletion events do not "cascade" down
through related models when using mass-deletion. So, while much slower,
it's necessary to iterate over the related models and call the delete
method on each, individually, so that the appropriate events fire.