tests: Fix a few lingering issues
Due to seeding-related changes, this test began failing and now
requires the CountriesSeeder to be run during setup.

Also, eliminated the potential for Faker to generate a username
that is already in-use (which would cause the test to fail) by using a
randomly-generated string with a much, much lower collision likelihood.

Finally, removed DB transaction that simply isn't necessary here.