chore(GrumPHP): Add database directory to ignore list
Note that if a file in this directory is staged to be commited it will
be scanned and will fail due to the namespace sniff. I'm not really
sure at this time how to fix/avoid this. To me it seems like it should
behave as we would expect it to. Possible bug ?