refact: Move password hashing from UserController to UserRepository
Given that password hashing isn't HTTP-specific (a password could
certainly be updated by other means, such as CLI), it shouldn't be done
in the controller. Instead, it should be done in the repository.

Note also that the username requirement has been removed from the
UpdateUser validation class because we have the user ID there, so the
username is unnecessary (unlike when creating).

Ref issue #57