refact(Controllers): Add abstract ApiController to cleanup controllers
This proposes that we add this abstract controller do reduce the code
duplication. Create new controllers on the fly by setting a few values
and moving on with life.

The compromise here is to not have dedicated requests like `StoreUser`
instead we set the rules on the controller, But I think the compromise
in the end is cleaner and makes things a bit more clear and together.

If we accept this proposal then I will merge it with the
`wip-api-controllers` branch and refactor the rest of the controllers to
conform to this new pattern.

Ref PR #71