tests: Correct issue arising from auto-incrementing during tests
We shouldn't assume anything about the values of auto-incremented
columns; there's no guarantee that a seeded table will have a record
with the ID of 1. Accordingly, this modifies the test such that
a new Genre is created and then used during model attachment.

This change also eliminates the need to seed the Genres table.