fix: Remove moniker field, as it is not needed here
Removing this prior to d8b00dd983bc25aba6feafc804caf46d5c57509d caused
a test failure; the offending test logic has been corrected.

To elaborate briefly as to why the the moniker field does not need to
be validated in this context: Artists are not permitted to change their
monikers in a self-service capacity, primarily because the moniker
is high-visibility and may e.g., end-up on the homepage, and
intentional profanity and so forth shouldn't be allowed to appear
without manual vetting. Furthermore, customers who have purchased a
given artist's music might be confused by a moniker change, so artists
shouldn't be allowed to change this value at a whim.