tests(Refact): Implement mechanism for making new model instances
This is a bare-minimum "proposal" as to one possible way of approaching
the need to have a means by which to make a new model instance in the
context of a Feature test.

To be clear, this commit implements *only* the bare-minimum changes
required to demonstrate the approach with passing tests, which is done
in the AlbumRepositoryTest's test_method_create_storesNewResource()
method.