refact: Rename "Cart" and "CartItem" to "Order" and "OrderItem"
Following an extensive side-channel discussion, we've decided that this
issue, the related issues, and the parent milestone are really more
about an "Order" than a "Shopping Cart".

Accordingly, we've renamed the above, as appropriate.

The fundamental distinction to be made here is that the Shopping Cart
will be handled entirely on the front-end, and all we're concerned with
here is handling its contents and baking them into an Order with
individual Order Items.

Ref issue #20