chore(Dependencies): Update dependencies
The GrumPHP upstream finally merged our changes so we can remove our
forked repository from the composer file. The features are due to
release in version `0.14.3` (currently 0.14.2) so we will use the master
branch on their repository until the official release.

In addition, I ran a `composer update` to update the composer.lock file
which includes the Laravel update to `5.7` from a previous commit.

Ref commit 6a774450846c12f4ff1875cb9fdbcb6cb1a910eb