feat(Api Routes): Add feature to exclude routes from default routes
In some cases an Api route may not require a certain route to be created
along with the default routes. Now with a new method `except($route)` we
can exclude an array of routes from the default list BEFORE adding the
default routes.

This feature prompted the need for this API to be more scalable thus
causing API changes. With this new approach we shouldn't worry about
the API breaking in the future.

New API Examples:

```php
$this->apiRoute($prefix, $controller)
    ->except(array $routes)
    // routes to exclude: 'index', 'show', 'store', 'update', 'destroy'

    ->addDefaultRoutes()
    // add default routes: 'index', 'show', 'store', 'update', 'destroy'

    ->addRoute(
      $uri,
      $controllerMethod,
      $httpMethod = 'get',
      $name = null
    );
    // if $name is null, route name will become the lowercase version
    // of $controllerMethod
```