tests(fix): Eliminate source of variability that was causing erratic results
The Album factory is used throughout the various test-cases, and
prior to this change, the is_active value was not being passed
explicitly, which was causing the value to be left to the factory's
implementation, which adds an element of randomness to the value.
Passing is_active=true, explicitly, in each factory call eliminates
the unwanted (random) behavior.

Also, there's no need to create three identical eligible Artists; only
one of each requirement type is necessary, so the extras were removed.