tests: Modify test because is_active column can no longer be used here
This change is a bit "hacky", as the repository to which this test
applies probably doesn't need Update capabilities at all, but this
test class extends a case in which tests for all major CRUD operations
are enforced via abstract methods.

The only "elegant" means by which to assemble CRUD tests that could
potentially omit any of the four CRUD operations while minimizing
repetition would be to compose them from traits. But I almost like this
unsightly hack better than implementing any such a la carte CRUD test
case.

And so, this remains as is, for now.