fix: Remove unused song_user table
Originally, this table's intention was to relate users with the
songs they had purchased, but as the project evolves, it's clear
that the mechanism by which these entities are related will be more
complex than this. Given that these changes do not break any tests,
all of this is safe to remove.