refact: Update Composer dependencies and refactor tests
The CountriesSeeder is better suited to reside in the indiehd/velkart
package, instead of here. As such, the vast majority of tests needed
to be refactored to account for this change.

Also, the User model was moved into a "models" subdirectory in Laravel
8, which was merged-in here from upstream, but this commit moves it
back into the "app" subdirectory, until all models can be moved into a
subdirectory.