tests: Make tests less brittle by removing hard-coded IDs
Historically, these tests always passed, but some recent change(s)
caused them to start failing. But, upon closer inspection, it seems
less than ideal to hard-code the IDs to 1, because ID-related
side-effects are particularly difficult to avoid in some situations.

In the interest of avoiding unwanted side-effects, this commit changes
the ID references to be dynamic, which allows the test to pass.